+++ custom_block/simple_with_meta.ts
@@ -1,5 +1,5 @@
 import { Block, KnownBlock } from '@slack/types';
-import { SummaryResults } from '..';
+import { SummaryResults } from '../src';
 
 export default function generateCustomLayoutSimpleMeta(
   summaryResults: SummaryResults,
